requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
# mediapipe is kept ONLY for Face Mesh blink/EAR detection (468-pt landmarks).
# Pose detection is handled by YOLOv8-pose (ultralytics) instead.
mediapipe>=0.10.0
ultralytics>=8.2.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
pyttsx3>=2.90
pynput>=1.7.6
pystray>=0.19.4
scikit-learn>=1.3.0
pygame>=2.5.0
